+++ chia/util/setup_nodes.py
@@ -8,7 +8,6 @@ from chia.full_node.full_node import FullNode
 from chia.full_node.full_node_api import FullNodeAPI
 from chia.protocols.shared_protocol import Capability
 from chia.server.server import ChiaServer
-from chia.server.start_data_layer import create_data_layer_service
 from chia.server.start_service import Service
 from chia.simulator.block_tools import BlockTools, create_block_tools_async, test_constants
 from chia.simulator.full_node_simulator import FullNodeSimulator
@@ -16,7 +15,7 @@ from chia.types.peer_info import PeerInfo
 from chia.util.hash import std_hash
 from chia.util.ints import uint16, uint32
 from chia.wallet.wallet_node import WalletNode
-from tests.setup_services import (
+from chia.util.setup_services import (
     setup_daemon,
     setup_farmer,
     setup_full_node,
@@ -28,7 +27,7 @@ from tests.setup_services import (
     setup_wallet_node,
 )
 from chia.simulator.time_out_assert import time_out_assert_custom_interval
-from tests.util.keyring import TempKeyring
+from chia.util.keyring import TempKeyring
 from chia.simulator.socket import find_available_listen_port
 
 
@@ -56,34 +55,6 @@ async def _teardown_nodes(node_aiters: List) -> None:
             pass
 
 
-async def setup_data_layer(local_bt):
-    # db_path = local_bt.root_path / f"{db_name}"
-    # if db_path.exists():
-    #     db_path.unlink()
-    config = local_bt.config["data_layer"]
-    # config["database_path"] = db_name
-    # if introducer_port is not None:
-    #     config["introducer_peer"]["host"] = self_hostname
-    #     config["introducer_peer"]["port"] = introducer_port
-    # else:
-    #     config["introducer_peer"] = None
-    # config["dns_servers"] = []
-    # config["rpc_port"] = port + 1000
-    # overrides = config["network_overrides"]["constants"][config["selected_network"]]
-    # updated_constants = consensus_constants.replace_str_to_bytes(**overrides)
-    # if simulator:
-    #     kwargs = service_kwargs_for_full_node_simulator(local_bt.root_path, config, local_bt)
-    # else:
-    #     kwargs = service_kwargs_for_full_node(local_bt.root_path, config, updated_constants)
-
-    service = create_data_layer_service(local_bt.root_path, config, connect_to_daemon=False)
-
-    await service.start()
-
-    yield service._api
-
-    service.stop()
-    await service.wait_closed()
 
 
 async def setup_two_nodes(consensus_constants: ConsensusConstants, db_version: int, self_hostname: str):
