+++ pyproject.toml
@@ -8,7 +8,7 @@ local_scheme = "no-local-version"
 
 [tool.black]
 line-length = 120
-target-version =  ['py37', 'py38', 'py39']
+target-version =  ['py37', 'py38']
 include = '''
 ^/(
     [^/]*.py
