+++ pytest.ini
@@ -1,9 +1,7 @@
 [pytest]
 ; logging options
-log_cli = False
-addopts = --verbose --tb=short -n auto -p no:monitor
-log_level = WARNING
-console_output_style = count
+log_cli=true
+log_level = INFO
 log_format = %(asctime)s %(name)s: %(levelname)s %(message)s
 asyncio_mode = strict
 markers =
