+++ tests/core/cmds/test_keys.py
@@ -12,7 +12,7 @@ from chia.util.keyring_wrapper import DEFAULT_KEYS_ROOT_PATH, KeyringWrapper, Le
 from click.testing import CliRunner, Result
 from keyring.backend import KeyringBackend
 from pathlib import Path
-from tests.util.keyring import TempKeyring
+from chia.util.keyring import TempKeyring
 from typing import Dict, List, Optional
