+++ tests/core/full_node/full_sync/test_full_sync.py
@@ -14,7 +14,7 @@ from chia.types.peer_info import PeerInfo
 from chia.util.hash import std_hash
 from chia.util.ints import uint16
 from tests.core.node_height import node_height_exactly, node_height_between
-from tests.setup_nodes import test_constants
+from chia.util.setup_nodes import test_constants
 from chia.simulator.time_out_assert import time_out_assert
 
 log = logging.getLogger(__name__)
