+++ tests/core/full_node/stores/test_full_node_store.py
@@ -23,9 +23,9 @@ from tests.blockchain.blockchain_test_utils import (
     _validate_and_add_block,
     _validate_and_add_block_no_error,
 )
-from tests.setup_nodes import test_constants as test_constants_original
+from chia.util.setup_nodes import test_constants as test_constants_original
 from tests.util.blockchain import create_blockchain
-from tests.util.keyring import TempKeyring
+from chia.util.keyring import TempKeyring
 
 
 test_constants = test_constants_original.replace(**{"DISCRIMINANT_SIZE_BITS": 32, "SUB_SLOT_ITERS_STARTING": 2**12})
