+++ tests/core/full_node/test_conditions.py
@@ -24,7 +24,7 @@ from chia.types.spend_bundle import SpendBundle
 from chia.util.errors import Err
 from chia.util.ints import uint32
 from chia.simulator.block_tools import BlockTools
-from tests.util.keyring import TempKeyring
+from chia.util.keyring import TempKeyring
 
 from .ram_db import create_ram_blockchain
 from ...blockchain.blockchain_test_utils import _validate_and_add_block
