+++ tests/core/full_node/test_full_node.py
@@ -49,7 +49,7 @@ from tests.core.full_node.stores.test_coin_store import get_future_reward_coins
 from tests.core.full_node.test_mempool_performance import wallet_height_at_least
 from tests.core.make_block_generator import make_spend_bundle
 from tests.core.node_height import node_height_at_least
-from tests.setup_nodes import test_constants
+from chia.util.setup_nodes import test_constants
 from chia.simulator.time_out_assert import time_out_assert, time_out_assert_custom_interval, time_out_messages
