+++ tests/core/server/test_rate_limits.py
@@ -13,7 +13,7 @@ from chia.server.ws_connection import WSChiaConnection
 from chia.types.peer_info import PeerInfo
 from chia.util.ints import uint16
 from tests.conftest import node_with_params
-from tests.setup_nodes import test_constants
+from chia.util.setup_nodes import test_constants
 
 
 constants = test_constants
