+++ tests/core/test_cost_calculation.py
@@ -18,7 +18,7 @@ from chia.types.blockchain_format.program import Program, SerializedProgram
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.generator_types import BlockGenerator
 from chia.wallet.puzzles import p2_delegated_puzzle_or_hidden_puzzle
-from tests.setup_nodes import test_constants
+from chia.util.setup_nodes import test_constants
 from tests.util.misc import assert_runtime
 
 from .make_block_generator import make_block_generator
