+++ tests/core/util/test_streamable.py
@@ -44,7 +44,7 @@ from chia.util.streamable import (
     streamable_from_dict,
     write_uint32,
 )
-from tests.setup_nodes import test_constants
+from chia.util.setup_nodes import test_constants
 
 
 def test_int_not_supported() -> None:
