+++ tests/wallet/rpc/test_dl_wallet_rpc.py
@@ -14,7 +14,7 @@ from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.peer_info import PeerInfo
 from chia.util.ints import uint16, uint32, uint64
 from chia.wallet.db_wallet.db_wallet_puzzles import create_mirror_puzzle
-from tests.setup_nodes import SimulatorsAndWalletsServices
+from chia.util.setup_nodes import SimulatorsAndWalletsServices
 from tests.util.rpc import validate_get_routes
 
 log = logging.getLogger(__name__)
