+++ tests/wallet/test_wallet_blockchain.py
@@ -8,7 +8,7 @@ from chia.types.weight_proof import WeightProof
 from chia.util.generator_tools import get_block_header
 from chia.wallet.key_val_store import KeyValStore
 from chia.wallet.wallet_blockchain import WalletBlockchain
-from tests.setup_nodes import test_constants
+from chia.util.setup_nodes import test_constants
 from tests.util.db_connection import DBConnection
